changes from today's lab

+++ boolean.h
@@ -9,6 +9,6 @@
 #define FALSE 0
 
 //TODO: determine how to track if init has already been called; ask Rachel if solution is acceptable? (Done and yes)
-typedef unsigned int boolean;
+typedef short int boolean;
 
 #endif //HW5_BOOLEAN_H
